requirements.txt
asgiref==3.8.1
awscli==1.37.7
boto3==1.36.7
botocore==1.36.7
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.5
django-storages==1.14.4
djangorestframework==3.16.0
docutils==0.16
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
transbank-sdk==6.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.8.2
